+++ app/store/client.js
@@ -1,7 +1,10 @@
 let mongodb = require("mongodb"),
     client = new mongodb.MongoClient(
         process.env.MONGODB_URL,
-        { useNewUrlParser: true }
+        {
+            useNewUrlParser: true,
+            useUnifiedTopology: true
+        }
     );
 
 module.exports = client;