+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## **1.0.4** - *2020-03-06*
 * node.js packages updated;
+* Rollup.js updated to v2;
 * `Dashboard/Gallery`:
     * pre-populate with XMP data fixed;
     * editing photo's photoSets fixed;